User can see build status. #25
